Added SLSC - stationary receiver
Added code for both integral and chunk-wise SLSC reconstruction (for a stationary receiver  only).

Also some updates on comments.

Furthermore, some changes to the MATLAB benchmark script, to limit number of repeat timings (only reconstructions faster than 1s will be repeated until at least 1s of compute time is spent).

Finally, slight adjustments to the chunk sizes for the three methods. For DAS, this resulted in a few (5-10)% speed up; for DMAS and SLSC about a factor of 1.5-2.